test: move api mocks to separate folders
